+++ lib/slkecho/option_parser.rb
@@ -2,13 +2,12 @@
 
 module Slkecho
   class OptionParser
-    def option_parser # rubocop:disable Metrics/AbcSize,Metrics/MethodLength
+    def option_parser # rubocop:disable Metrics/AbcSize
       @option_parser ||= ::OptionParser.new do |o|
         o.banner = "Usage: slkecho [options] message"
         o.program_name = "slkecho"
         o.version = Slkecho::VERSION
         o.on("-c", "--channel CHANNEL", "Slack channel to post the message") { @options.channel = _1 }
-        o.on("-s", "--subject SUBJECT", "Subject of message") { @options.subject = _1 }
         o.on("-m", "--mention EMAIL", "Mention to user by email") { @options.mention = _1 }
         o.on("--username USERNAME", "Set your bot's user name") { @options.username = _1 }
         o.on("--icon-url ICON_URL", "URL to an image to use as the icon for this message") { @options.icon_url = _1 }